# perf: faster reading of files with many branches