This PR fixes #1442 (at least partially). It's not clear to me why this getter is needed, but it has a huge performance cost, so I switched to using the branch names directly.

This change gives it a huge boost in performance for trees with many branches, although it is still significantly slower than Uproot3.

For comparison, I generated a file with 1000 branches, as done in #1442 and the output of
```python
%timeit uproot.open("test.root:df").arrays(library="pd")
```
went from
```
3.07 s ± 46.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
to
```
353 ms ± 7.82 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

@ianna do you know why this getter could have been needed in the first place? All the tests seem to pass, so I'm not sure why it was needed.