@ikrommyd 2025-06-11
Ci is failing due to numpy 2.3 and awkward. Perhaps we need a quick awkward patch release.