@ariostas 2025-06-11
I just found out that another big improvement could be made by changing the way `cache_key` works. Currently, it uses the index, which is very slow to find for TTrees with many branches, so we should think of a better alternative.

https://github.com/scikit-hep/uproot5/blob/61835f10456541965ec7fc6ba37ddb07ecb5cfe9/src/uproot/behaviors/TBranch.py#L1960-L1984