@ariostas 2025-06-11
After the last commit it's actually back to the speed of Uproot3!